pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kasparro-agentic"
version = "2.0.0"
description = "LangGraph-powered multi-agent content generation system"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Saad Ilkal", email = "saad@example.com"}
]
keywords = ["langgraph", "langchain", "agents", "content-generation", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.1.0",
    "langchain-ollama>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]
openai = [
    "langchain-openai>=0.0.5",
]

[project.scripts]
run-pipeline = "agents.graph:run_pipeline"

[project.urls]
"Homepage" = "https://github.com/Fatal777/kasparro-agentic"
"Bug Tracker" = "https://github.com/Fatal777/kasparro-agentic/issues"

[tool.setuptools.packages.find]
include = ["agents*", "core*", "logic_blocks*", "api*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
